remove comment from before commits
2016-02-18 19:21 Moved to CustomScanner in order to keep the TicTacToe
class cleaner. Found the String trim() method

2016-02-18 17:25 I initially only used split() and did not use
replaceAll(). I changed to the using replaceAll() because it allows to
disregard any blank space that the user might input and only focus on
the length of the user's input (should be 2, [character, number]) and
the values of the input, although that is not validated here. I no
longer check the length of each split input and the length of the input
array. I found this optimization when I put a space in front of input
before my a2 (" a2") and it returned an unexpected error.